+++ lib/surface/constructs/for_construct.ex
@@ -1,0 +1,3 @@
+defmodule Surface.Constructs.For do
+  use Surface.Construct
+end
